changing and fixing for a better website
